+++ lib/rubocop/cli.rb
@@ -53,6 +53,7 @@ module RuboCop
 
       ConfigLoader.debug = @options[:debug]
       ConfigLoader.auto_gen_config = @options[:auto_gen_config]
+      ConfigLoader.exclude_limit = @options[:exclude_limit]
 
       @config_store.options_config = @options[:config] if @options[:config]